Trait pallet_liquidity_mining::pallet::Config

pub trait Config<I: 'static = ()>: Config + TypeInfo {

    type Event: From<Event<Self, I>> + IsType<<Self as Config>::Event>;
    type AssetId: Parameter + Member + Copy + MaybeSerializeDeserialize + MaxEncodedLen + Into<u32>;
    type MultiCurrency: MultiCurrency<Self::AccountId, CurrencyId = Self::AssetId, Balance = Balance>;
    type PalletId: Get<PalletId>;
    type MinTotalFarmRewards: Get<Balance>;
    type MinPlannedYieldingPeriods: Get<Self::BlockNumber>;
    type BlockNumberProvider: BlockNumberProvider<BlockNumber = Self::BlockNumber>;
    type AmmPoolId: Parameter + Member + Clone + FullCodec + MaxEncodedLen;
    type MaxFarmEntriesPerDeposit: Get<u32>;
    type MaxYieldFarmsPerGlobalFarm: Get<u32>;
    type AssetRegistry: Registry<Self::AssetId, Vec<u8>, Balance, DispatchError> + GetByKey<Self::AssetId, Balance>;
    type NonDustableWhitelistHandler: DustRemovalAccountWhitelist<Self::AccountId, Error = DispatchError>;
    type PriceAdjustment: PriceAdjustment<GlobalFarmData<Self, I>, Error = DispatchError, PriceAdjustment = FixedU128>;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self, I>> + IsType<<Self as Config>::Event>

type AssetId: Parameter + Member + Copy + MaybeSerializeDeserialize + MaxEncodedLen + Into<u32>

Asset type.

type MultiCurrency: MultiCurrency<Self::AccountId, CurrencyId = Self::AssetId, Balance = Balance>

Currency for transfers.

type PalletId: Get<PalletId>

Pallet id.

type MinTotalFarmRewards: Get<Balance>

Minimum total rewards to distribute from global farm during liquidity mining.

type MinPlannedYieldingPeriods: Get<Self::BlockNumber>

Minimum number of periods to run liquidity mining program.

type BlockNumberProvider: BlockNumberProvider<BlockNumber = Self::BlockNumber>

The block number provider

type AmmPoolId: Parameter + Member + Clone + FullCodec + MaxEncodedLen

Id used to identify amm pool in liquidity mining pallet.

type MaxFarmEntriesPerDeposit: Get<u32>

Maximum number of yield farms same LP shares can be re/deposited into. This value always MUST BE >= 1.

type MaxYieldFarmsPerGlobalFarm: Get<u32>

Max number of yield farms can exist in global farm. This includes all farms in the storage(active, stopped, terminated).

type AssetRegistry: Registry<Self::AssetId, Vec<u8>, Balance, DispatchError> + GetByKey<Self::AssetId, Balance>

Asset Registry - used to check if asset is correctly registered in asset registry and provides information about existential deposit of the asset.

type NonDustableWhitelistHandler: DustRemovalAccountWhitelist<Self::AccountId, Error = DispatchError>

Account whitelist manager to exclude pool accounts from dusting mechanism.

type PriceAdjustment: PriceAdjustment<GlobalFarmData<Self, I>, Error = DispatchError, PriceAdjustment = FixedU128>