Crate pallet_liquidity_mining

Liquidity mining pallet

Overview

This pallet provide functionality for liquidity mining program with time incentive(loyalty factor). Users are rewarded for each period they stay in liq. mining program.

Reward per one period is derived from the user’s loyalty factor which grows with time(periods) the user is in the liq. mining and amount of LP shares user locked into deposit. User’s loyalty factor is reset if the user exits and reenters liquidity mining pool. User can claim rewards without resetting loyalty factor, only withdrawing shares is penalized by loyalty factor reset. User is rewarded from the next period after he enters.

User deposit in liquidity mining pool is represented by an NFT which is minted for the user when he enters liq. mining and is burned when he exits. NFT representing deposit is tradable.

Re-exports

pub use pallet::*;

Modules

migration

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_liquidity_mining

Structs

Deposit

GlobalPool

This struct represents the state a of single liquidity mining program. LiquidityPoolYieldFarms are rewarded from GlobalPool based on their stake in GlobalPool. LiquidityPoolYieldFarm stake in GlobalPool is derived from users stake in LiquidityPoolYieldFarm.

LiquidityPoolYieldFarm

LoyaltyCurve

Constants

MAX_NFT_INSTANCE_SEQUENCER

NFT_SEQUENCE_BYTES

POOL_ID_BYTES

Type Definitions

AccountIdOf

AssetIdOf

BlockNumberFor

GlobalPoolId

NftClassIdOf

NftInstanceIdOf

PeriodOf

PoolId

PoolMultiplier